mvp/requirements.txt
schedule==0.5.0
python-crontab==2.3.5
paramiko==2.4.2
ssh-authorizer==1.3
boto3==1.9.33